+++ Sources/Spritz/Models/Models.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public extension Spritz {
+  /// Namespace for model objects.
+  enum Models {}
+}